impr(settings): match custom settings by their options in search (@OmTiwariPandey)
